+++ tests/testthat/test_NMextractText.R
@@ -7,7 +7,7 @@ context("NMextractText")
 test_that("basic",{
 
     fileRef <- "testReference/NMextractText_1.rds"
-    file.lst <- NMdata_filepath("examples/nonmem/xgxr004.lst")
+    file.lst <- system.file("examples/nonmem/xgxr004.lst",package="NMdata")
     res1 <- NMextractText(file.lst,section="THETA",char.section="\\$")
 
     expect_equal_to_reference(res1,fileRef,version=2)